requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.2
pydantic-settings==2.6.1
psutil==5.9.8
aiofiles==25.1.0

# Database and Vector Store
chromadb==0.4.18
sqlalchemy==2.0.36
alembic==1.12.1
asyncpg==0.29.0

# AI/ML
ollama==0.1.7
langchain==0.0.350
langchain-community==0.0.10
sentence-transformers==2.2.2
torch>=2.5.0
transformers>=4.36.2

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
# pyyubico==1.5.3  # Optional - for YubiKey support

# Email Processing
imaplib2==3.6
email-validator==2.1.1
python-magic==0.4.27

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
opencv-python==4.8.1.78
pymupdf==1.23.8

# Voice Processing
SpeechRecognition==3.10.1
pyttsx3==2.98
pyaudio==0.2.14
Jinja2==3.1.2

# Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3
schedule==1.2.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0